Re-structured functions in the package. The pikepdf package should only be necesary for the CLI.
